polished event services

+++ Backend/309_Server/src/main/java/callyourmom/repository/EventRepository.java
@@ -4,6 +4,6 @@ import callyourmom.user.Event;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface EventRepository extends JpaRepository<Event, Long> {
+public interface EventRepository extends JpaRepository<Event, Integer> {
     Event findByTitle(String title);
-}
+}
